refactor: it split up tests context.
