+++ pkg/arango_dialector_test.go
@@ -9,12 +9,6 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-type User struct {
-	gorm.Model
-	Name  string
-	Email string
-}
-
 var _ = Describe("ArangoDB Dialector", func() {
 	arangoDBTestConfig := newArangoDBTestConfig()
 
@@ -40,28 +34,6 @@ var _ = Describe("ArangoDB Dialector", func() {
 		It("creates a collection", func() {
 			err := db.AutoMigrate(&User{})
 			Expect(err).NotTo(HaveOccurred())
-		})
-
-		It("inserts a record into the collection", func() {
-			err := db.AutoMigrate(&User{})
-			Expect(err).NotTo(HaveOccurred())
-
-			user := &User{
-				Name:  "Joselito",
-				Email: "joselitofilhoo@gmail.com",
-			}
-			tx := db.Create(user)
-			Expect(tx).NotTo(BeNil())
-			Expect(tx.Error).To(BeNil())
-			Expect(tx.Statement.Dest).NotTo(BeNil())
-
-			newUser := tx.Statement.Dest.(*User)
-			Expect(newUser.ID).NotTo(BeZero())
-			Expect(newUser.Name).To(Equal(user.Name))
-			Expect(newUser.Email).To(Equal(user.Email))
-			Expect(newUser.CreatedAt).NotTo(BeZero())
-			Expect(newUser.UpdatedAt).NotTo(BeZero())
-			Expect(newUser.DeletedAt).To(BeZero())
 		})
 	})
 })
